+++ demo/platformer.js
@@ -189,7 +189,7 @@ scene("game", ({ levelId, coins } = { levelId: 0, coins: 0 }) => {
 		}
 	})
 
-	player.onCollisionResolve(() => {
+	player.onPhysicsResolve(() => {
 		// Set the viewport center to player.pos
 		camPos(player.pos)
 	})