+++ src/kaboom.ts
@@ -3974,7 +3974,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 					cleanups.push(this.onHover(() => setCursor(this.area.cursor)))
 				}
 
-				cleanups.push(this.onCollisionActive((obj, col) => {
+				cleanups.push(this.onCollideUpdate((obj, col) => {
 					if (!this.colliding[obj.id]) {
 						this.trigger("collide", obj, col)
 					}
@@ -3988,7 +3988,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 					const col = this.colliding[id]
 					if (!this.checkCollision(col.target)) {
 						delete this.colliding[id]
-						this.trigger("collisionEnd", col.target, col)
+						this.trigger("collideEnd", col.target, col)
 					}
 				}
 			},
@@ -4111,25 +4111,25 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 				}
 			},
 
-			onCollisionActive(
+			onCollideUpdate(
 				tag: Tag | ((obj: GameObj, col?: Collision) => void),
 				cb?: (obj: GameObj, col?: Collision) => void,
 			): EventCanceller {
 				if (typeof tag === "function" && cb === undefined) {
-					return this.on("collisionActive", tag)
+					return this.on("collideUpdate", tag)
 				} else if (typeof tag === "string") {
-					return this.on("collisionActive", (obj, col) => obj.is(tag) && cb(obj, col))
+					return this.on("collideUpdate", (obj, col) => obj.is(tag) && cb(obj, col))
 				}
 			},
 
-			onCollisionEnd(
+			onCollideEnd(
 				tag: Tag | ((obj: GameObj) => void),
 				cb?: (obj: GameObj) => void,
 			): EventCanceller {
 				if (typeof tag === "function" && cb === undefined) {
-					return this.on("collisionEnd", tag)
+					return this.on("collideEnd", tag)
 				} else if (typeof tag === "string") {
-					return this.on("collisionEnd", (obj) => obj.is(tag) && cb(obj))
+					return this.on("collideEnd", (obj) => obj.is(tag) && cb(obj))
 				}
 			},
 
@@ -4615,7 +4615,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 				// static vs static: don't resolve
 				// static vs non-static: always resolve non-static
 				// non-static vs non-static: resolve the first one
-				cleanups.push(this.onCollisionActive((other, col) => {
+				cleanups.push(this.onCollideUpdate((other, col) => {
 
 					if (!other.is("body")) {
 						return
@@ -4637,12 +4637,12 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 					col2.source.pos = col2.source.pos.add(col2.displacement)
 					// TODO: update all children transform?
 					col2.source.transform = calcTransform(col2.source)
-					col2.source.trigger("collisionResolve", col2)
-					col2.target.trigger("collisionResolve", col2.reverse())
+					col2.source.trigger("physicsResolve", col2)
+					col2.target.trigger("physicsResolve", col2.reverse())
 
 				}))
 
-				cleanups.push(this.onCollisionResolve((col) => {
+				cleanups.push(this.onPhysicsResolve((col) => {
 					if (game.gravity) {
 						if (col.isBottom() && this.isFalling()) {
 							velY = 0
@@ -4710,8 +4710,8 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 				cleanups.forEach((f) => f())
 			},
 
-			onCollisionResolve(action) {
-				return this.on("collisionResolve", action)
+			onPhysicsResolve(action) {
+				return this.on("physicsResolve", action)
 			},
 
 			curPlatform(): GameObj | null {
@@ -4868,7 +4868,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 			if (!events[state]) {
 				events[state] = {
 					enter: new Event(),
-					leave: new Event(),
+					end: new Event(),
 					update: new Event(),
 					draw: new Event(),
 				}
@@ -4919,7 +4919,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 
 				}
 
-				trigger("leave", oldState, ...args)
+				trigger("end", oldState, ...args)
 				this.state = state
 				trigger("enter", state, ...args)
 				trigger("enter", `${oldState} -> ${state}`, ...args)
@@ -4942,8 +4942,8 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 				return on("draw", state, action)
 			},
 
-			onStateLeave(state: string, action: () => void): EventCanceller {
-				return on("leave", state, action)
+			onStateEnd(state: string, action: () => void): EventCanceller {
+				return on("end", state, action)
 			},
 
 			update() {
@@ -5443,11 +5443,11 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 								if (res && !res.isZero()) {
 									// TODO: rehash if the object position is changed after resolution?
 									const col1 = new Collision(aobj, other, res)
-									aobj.trigger("collisionActive", other, col1)
+									aobj.trigger("collideUpdate", other, col1)
 									const col2 = col1.reverse()
 									// resolution only has to happen once
 									col2.resolved = col1.resolved
-									other.trigger("collisionActive", aobj, col2)
+									other.trigger("collideUpdate", aobj, col2)
 								}
 								checked.add(other.id)
 							}